main commit
Model Training and Evaluation: Implemented the training of an EfficientNet-B0 model for detecting the presence of St. George in images. The model was fine-tuned using transfer learning with pre-trained weights to improve performance on the given task.

Performance Metrics: Achieved high accuracy and balanced metrics, with precision, recall, and F1-scores demonstrating robust performance. Evaluated the model using classification reports and confusion matrices to assess its effectiveness.

Inference: Developed functionality to classify new images and determine the presence of St. George based on the trained model.